feat Question, reply 수정 삭제시 본인인증

+++ server/src/main/java/com/codestates/StackOverFlowClone/question/service/QuestionService.java
@@ -1,6 +1,7 @@
 package com.codestates.StackOverFlowClone.question.service;
 
 
+import com.codestates.StackOverFlowClone.member.service.MemberService;
 import com.codestates.StackOverFlowClone.question.entity.Question;
 import com.codestates.StackOverFlowClone.question.repository.QuestionRepository;
 import com.codestates.StackOverFlowClone.reply.entity.Reply;
@@ -8,6 +9,7 @@ import com.codestates.StackOverFlowClone.utils.CustomBeanUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -19,10 +21,13 @@ public class QuestionService {
 
     private final QuestionRepository questionRepository;
 
+    private final MemberService memberService;
+
     private final CustomBeanUtils<Question> beanUtils;
 
-    public QuestionService(QuestionRepository questionRepository, CustomBeanUtils<Question> beanUtils){
+    public QuestionService(QuestionRepository questionRepository, MemberService memberService, CustomBeanUtils<Question> beanUtils){
         this.questionRepository = questionRepository;
+        this.memberService = memberService;
         this.beanUtils = beanUtils;
     }
 
@@ -38,6 +43,12 @@ public class QuestionService {
 
         Question updatingQuestion =
                 beanUtils.copyNonNullProperties(question, findQuestion);
+
+        // 본인만 게시물 수정
+        long tokenMemberId = memberService.findTokenMemberId();
+        if(!isAuthenticatedMember(tokenMemberId, updatingQuestion.getMember().getMemberId())){
+            throw new RuntimeException();
+        }
 
         return questionRepository.save(updatingQuestion);
     }
@@ -79,6 +90,18 @@ public class QuestionService {
 
         Question findQuestion = findQuestion(questionId);
 
+        // 본인만 게시물 삭제
+        long tokenMemberId = memberService.findTokenMemberId();
+        if(!isAuthenticatedMember(tokenMemberId, findQuestion.getMember().getMemberId())){
+            throw new RuntimeException();
+        }
+
         questionRepository.delete(findQuestion);
     }
+
+    public boolean isAuthenticatedMember(long tokenMemberId, long questionId){
+
+        return tokenMemberId == questionId;
+    }
+
 }
